Move DetailMatchResult into its own class
