+++ presto-main/src/test/java/com/facebook/presto/sql/planner/assertions/AggregationMatcher.java
@@ -25,6 +25,8 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import static com.facebook.presto.sql.planner.assertions.DetailMatchResult.NO_MATCH;
+import static com.facebook.presto.sql.planner.assertions.DetailMatchResult.match;
 import static com.google.common.base.MoreObjects.toStringHelper;
 import static com.google.common.base.Preconditions.checkState;
 
@@ -55,11 +57,11 @@ public class AggregationMatcher
         AggregationNode aggregationNode = (AggregationNode) node;
 
         if (groupId.isPresent() != aggregationNode.getGroupIdSymbol().isPresent()) {
-            return DetailMatchResult.NO_MATCH;
+            return NO_MATCH;
         }
 
         if (groupingSets.size() != aggregationNode.getGroupingSets().size()) {
-            return DetailMatchResult.NO_MATCH;
+            return NO_MATCH;
         }
 
         List<Symbol> aggregationsWithMask = aggregationNode.getAggregations()
@@ -70,22 +72,22 @@ public class AggregationMatcher
                 .collect(Collectors.toList());
 
         if (aggregationsWithMask.size() != masks.keySet().size()) {
-            return DetailMatchResult.NO_MATCH;
+            return NO_MATCH;
         }
 
         for (Symbol symbol : aggregationsWithMask) {
             if (!masks.keySet().contains(symbol)) {
-                return DetailMatchResult.NO_MATCH;
+                return NO_MATCH;
             }
         }
 
         for (int i = 0; i < groupingSets.size(); i++) {
             if (!matches(groupingSets.get(i), aggregationNode.getGroupingSets().get(i))) {
-                return DetailMatchResult.NO_MATCH;
+                return NO_MATCH;
             }
         }
 
-        return DetailMatchResult.match();
+        return match();
     }
 
     static <T> boolean matches(Collection<T> expected, Collection<T> actual)
