+++ presto-main/src/test/java/com/facebook/presto/sql/planner/assertions/GroupIdMatcher.java
@@ -22,8 +22,8 @@ import com.facebook.presto.sql.planner.plan.PlanNode;
 import java.util.List;
 import java.util.Map;
 
-import static com.facebook.presto.sql.planner.assertions.Matcher.DetailMatchResult.NO_MATCH;
-import static com.facebook.presto.sql.planner.assertions.Matcher.DetailMatchResult.match;
+import static com.facebook.presto.sql.planner.assertions.DetailMatchResult.NO_MATCH;
+import static com.facebook.presto.sql.planner.assertions.DetailMatchResult.match;
 import static com.google.common.base.MoreObjects.toStringHelper;
 import static com.google.common.base.Preconditions.checkState;
